[Fix] Renamed directions
