+++ src/classes/TableColumn.py
